+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - Added `BarSeries#copy(String name, int maximumBarCount)`
 - Added `BarSeries#addLastBar(Bar bar)`
 - Added `BarSeries#replaceBarsIfChanged(BarSeries barSeries, List<Bar> newBars)`
+- Added `BarSeries#getSeriesPeriod(ChronoUnit unit)`
 
 
 ## 0.18 (released May 15, 2025)